requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=5.28.0
requests==2.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
